Add job kill/del

+++ internal/server/job.go
@@ -2,6 +2,8 @@ package server
 
 import (
 	"context"
+	"errors"
+	"fmt"
 	"syscall"
 
 	"github.com/cedana/cedana/pkg/api/daemon"
@@ -25,6 +27,12 @@ func (s *Server) Kill(ctx context.Context, req *daemon.KillReq) (*daemon.KillRes
 		return nil, status.Errorf(codes.Internal, "failed to list jobs: %v", err)
 	}
 
+	if len(jobs) == 0 {
+		return nil, status.Errorf(codes.NotFound, "no jobs found")
+	}
+
+	errs := []error{}
+
 	for _, job := range jobs {
 		if job.GetProcess().GetInfo().GetIsRunning() {
 			pid := job.GetProcess().GetPID()
@@ -32,10 +40,11 @@ func (s *Server) Kill(ctx context.Context, req *daemon.KillReq) (*daemon.KillRes
 			if err != nil {
 				log.Error().Err(err).Msgf("failed to kill job %s", job.GetJID())
 			}
+			errs = append(errs, err)
 		}
 	}
 
-	return &daemon.KillResp{}, err
+	return &daemon.KillResp{}, errors.Join(errs...)
 }
 
 func (s *Server) Delete(ctx context.Context, req *daemon.DeleteReq) (*daemon.DeleteResp, error) {
@@ -44,17 +53,24 @@ func (s *Server) Delete(ctx context.Context, req *daemon.DeleteReq) (*daemon.Del
 		return nil, status.Errorf(codes.Internal, "failed to list jobs: %v", err)
 	}
 
+	if len(jobs) == 0 {
+		return nil, status.Errorf(codes.NotFound, "no jobs found")
+	}
+
+	errs := []error{}
+
 	for _, job := range jobs {
 		// Don't delete running jobs
 		if job.GetProcess().GetInfo().GetIsRunning() {
-			err = status.Errorf(codes.FailedPrecondition, "job %s is running", job.GetJID())
+			errs = append(errs, fmt.Errorf("job %s is running", job.GetJID()))
 			continue
 		}
 		err = s.db.DeleteJob(ctx, job.GetJID())
 		if err != nil {
 			log.Error().Err(err).Msgf("failed to delete job %s", job.GetJID())
 		}
+		errs = append(errs, err)
 	}
 
-	return &daemon.DeleteResp{}, err
+	return &daemon.DeleteResp{}, errors.Join(errs...)
 }
